requirements.txt
pandas==2.1.1
termcolor==2.3.0
python-dotenv==1.0.0
